requirements.txt
absl-py==0.9.0
blis==0.4.1
boto==2.49.0
boto3==1.13.19
botocore==1.16.19
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
contractions==0.0.25
cymem==2.0.3
docutils==0.15.2
en-core-web-lg==2.2.5
Flask==1.1.2
future==0.18.2
gensim==3.8.3
google-auth==1.16.0
google-auth-oauthlib==0.4.1
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
Markdown==3.2.2
MarkupSafe==1.1.1
murmurhash==1.0.2
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.4
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
psycopg2-binary==2.8.5
pyahocorasick==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
sentencepiece==0.1.91
six==1.15.0
smart-open==2.0.0
spacey==0.1.1
spacy==2.2.4
SQLAlchemy==1.3.17
srsly==1.0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
textsearch==0.0.17
thinc==7.4.0
torch==1.5.0
torchtext==0.6.0
tqdm==4.46.0
Unidecode==1.1.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
